Wrap all accordion components and use isExpanded
